+++ RabbitMQ.Stream.Client/AMQP/FormatCode.cs
@@ -92,29 +92,5 @@ namespace RabbitMQ.Stream.Client.AMQP
             UUID = 0x98; // UUID as defined in section 4.1.2 of RFC-4122
     }
 
-    public static class DataCode
-    {
-        public const int Size = 3;
-
-        public static byte Read(ReadOnlySequence<byte> amqpData)
-        {
-            var offset = WireFormatting.ReadByte(amqpData, out var marker);
-            offset += WireFormatting.ReadByte(amqpData.Slice(offset), out var descriptor);
-            WireFormatting.ReadByte(amqpData.Slice(offset), out var value);
-            return value;
-        }
-
-        public static int Write(Span<byte> span, byte data)
-        {
-            var offset = WireFormatting.WriteByte(span, FormatCode.Described);
-            offset += WireFormatting.WriteByte(span.Slice(offset), FormatCode.SmallUlong);
-            offset += WireFormatting.WriteByte(span.Slice(offset), data);
-            return offset;
-        }
-        
-        public const byte ApplicationData = 0x75;
-        public const byte MessageAnnotations = 0x72;
-        public const byte MessageProperties = 0x73;
-        public const byte ApplicationProperties = 0x74;
-    }
+    
 }
